pyproject.toml
[project]
name = "atprotobin"
version = "1.1.1"
description = "ATProto based pastebin"
readme = {file = "README.md", content-type = "text/markdown"}
authors = [
    { name = "Public Domain", email = "publicdomainrelay@protonmail.com" }
]
license = {text = "Unlicense"}
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.55",
    "uvicorn[standard]>=0.32.1",
    "fastapi[standard]>=0.115.5",
    "pillow>=11.0.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.19",
    "pygments>=2.18.0",
    "pydantic>=2.10.2",
    "markdown2>=2.5.1",
]

[project.urls]
Repository = "https://github.com/publicdomainrelay/atprotobin.git"
Issues = "https://github.com/publicdomainrelay/atprotobin/issues"

[project.scripts]
atprotobin = "atprotobin.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
